Increase depth bit size in QBVH rebalance.
Parry occasionally panics due to add overflow (#146).  I couldn't write
a test that reliably replicates the condition, but since it occurs at
`src/partitioning/update.rs:367` and the only addition on that line is
`depth + 1`, I'm willing to bet money that the problem is that we are
overflowing `depth`, which is a `u8`.  I bumped it to `u32`, which
should still have low memory cost and be more space than we ever need
for a tree structure.